Added nmap checking and installing.
I implemented the function that checks to see if nmap is installed and if it is not, prompt the user to install it. If they answer "Y" or "YES", the program will attempt to run sudo apt-get install python-nmap. If the process is unsuccessful, an error is displayed and asked if they would like to try again.