Added checking for nmap
Sorry I am having to upload this as all one change. I can't get git to do a request-pull (Most likely I did it wrong, lol).

I added a function to check to make sure nmap is installed and renamed check_for_nmap() to check for python nmap. For some reason, I was thinking python-nmap was a python version of nmap, not a dependency, lol. If nmap is not installed, it prompts the user to install it for them.